+++ src/app.js
@@ -1,9 +1,3 @@
-Object.assign = Object.assign && typeof Object.assign === 'function' ? Object.assign : function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key] } } } return target }
-Array.from = Array.from && typeof Array.from === 'function' ? Array.from : obj => [].slice.call(obj)
-
-import Wux from 'components/wux'
-import WxValidate from 'assets/plugins/WxValidate'
-
 App({
 	onLaunch() {
 		console.log('onLaunch')
@@ -14,6 +8,4 @@ App({
 	onHide() {
 		console.log('onHide')
 	},
-	Wux: Wux, 
-	WxValidate: (rules, messages) => new WxValidate(rules, messages), 
 })