+++ src/pages/barcode/index.js
@@ -1,13 +1,11 @@
-const App = getApp()
+import { $wuxBarcode } from '../../components/wux'
 
 Page({
 	data: {
 		value: '', 
 	},
 	onLoad() {
-		this.$wuxBarcode = App.Wux().$wuxBarcode
-
-		this.$wuxBarcode('barcode', '9787115335500')
+		new $wuxBarcode('barcode', '9787115335500')
 	},
 	bindinput(e) {
 		const value = e.detail.value
@@ -20,6 +18,6 @@ Page({
 			value, 
 		})
 
-		this.$wuxBarcode('barcode', value)
+		new $wuxBarcode('barcode', value)
 	},
 })