+++ src/pages/countdown/index.js
@@ -1,12 +1,11 @@
-const App = getApp()
+import { $wuxCountDown } from '../../components/wux'
 
 Page({
 	data: {},
 	onLoad() {
 		const that = this
-		that.$wuxCountDown = App.Wux().$wuxCountDown
-
-		that.c1 = that.$wuxCountDown({
+		
+		that.c1 = new $wuxCountDown({
 			date: 'June 7, 2087 15:03:25', 
 			render(date) {
 				const years = this.leadingZeros(date.years, 4)  + ' 年 '
@@ -21,7 +20,7 @@ Page({
 			}, 
 		})
 
-		that.c3 = that.$wuxCountDown({
+		that.c3 = new $wuxCountDown({
 			date: +(new Date) + 60000 * 20, 
 			render(date) {
 				const min = this.leadingZeros(date.min, 2)  + ' 分 '
@@ -36,7 +35,7 @@ Page({
 	vcode() {
 		const that = this
 		if (that.c2 && that.c2.interval) return !1
-		that.c2 = that.$wuxCountDown({
+		that.c2 = new $wuxCountDown({
 			date: +(new Date) + 60000, 
 			onEnd() {
 				that.setData({