+++ src/pages/prompt/index.js
@@ -1,4 +1,4 @@
-const App = getApp()
+import { $wuxPrompt } from '../../components/wux'
 const sliderWidth = 96
 
 Page({	
@@ -9,13 +9,12 @@ Page({
         sliderLeft: 0
     },
     onLoad() {
-		this.$wuxPrompt = App.Wux().$wuxPrompt
-		this.$wuxPrompt.init('msg1', {
+		$wuxPrompt.init('msg1', {
 			title: '空空如也', 
 			text: '暂时没有相关数据', 
 		}).show()
 
-		this.$wuxPrompt.init('msg2', {
+		$wuxPrompt.init('msg2', {
 			icon: '../../assets/images/iconfont-order.png', 
 			title: '您还没有相关的订单', 
 			text: '可以去看看有哪些想买', 
@@ -29,7 +28,7 @@ Page({
 			},
 		}).show()
 
-		this.$wuxPrompt.init('msg3', {
+		$wuxPrompt.init('msg3', {
 			icon: '../../assets/images/iconfont-empty.png', 
 			title: '暂无待评价订单', 
 		}).show()