+++ src/pages/toast/index.js
@@ -1,12 +1,10 @@
-const App = getApp()
+import { $wuxToast } from '../../components/wux'
 
 Page({
 	data: {},
-	onLoad() {
-		this.$wuxToast = App.Wux().$wuxToast
-	},
+	onLoad() {},
 	showToast() {
-		this.$wuxToast.show({
+		$wuxToast.show({
 			type: 'success',
 			timer: 1500,
 			color: '#fff',
@@ -15,7 +13,7 @@ Page({
 		})
 	},
 	showToastCancel() {
-		this.$wuxToast.show({
+		$wuxToast.show({
 			type: 'cancel',
 			timer: 1500,
 			color: '#fff',
@@ -24,7 +22,7 @@ Page({
 		})
 	},
 	showToastErr() {
-		this.$wuxToast.show({
+		$wuxToast.show({
 			type: 'forbidden',
 			timer: 1500,
 			color: '#fff',
@@ -33,7 +31,7 @@ Page({
 		})
 	},
 	showToastText() {
-		this.$wuxToast.show({
+		$wuxToast.show({
 			type: 'text',
 			timer: 1500,
 			color: '#fff',